requirements.txt
Send2Trash>=1.5.0
empy>=3.3.4
pyyaml>=5.1.0
